feat: improve database stability, MDF workflows, and diacritic-insensitive search
- Enhanced database recovery by stabilizing TCP enablement with fallback to Unix socket and consolidating extension management with retries.
- Refined migration processes, enforcing versioning standards and prioritizing ORM-based seeding to prevent schema mismatches.
- Introduced diacritic- and quote-insensitive search by normalizing terms and queries for prefix match scalability.
- Updated `UploadService` to streamline language handling, align MDF tag logic, and improve record sorting with normalized lexemes.
- Enhanced upload review UX with server-side filtering, pagination, and persistent user preferences.
- Improved docs by purging deprecated features like trigram indexing and aligning deployment guidelines with compatibility constraints.

+++ src/database/models/meta.py
@@ -1,5 +1,5 @@
 # <!-- CRITICAL: NO EDITS WITHOUT APPROVED PLAN (Wait for "Go", "Proceed", or "Approved") -->
-from sqlalchemy import Column, Integer, String, TIMESTAMP, Text
+from sqlalchemy import Column, Integer, String, TIMESTAMP, Text, BigInteger
 from sqlalchemy.sql import func
 from ..base import Base
 
@@ -10,6 +10,6 @@ class SchemaVersion(Base):
     """
     __tablename__ = 'schema_version'
     id = Column(Integer, primary_key=True, autoincrement=True)
-    version = Column(Integer, nullable=False)
+    version = Column(BigInteger, nullable=False)
     applied_at = Column(TIMESTAMP(timezone=True), server_default=func.now())
     description = Column(Text)
